pyproject.toml
[project]
name = "playlist-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Anthony Kovach", email = "aj@ajxd2.dev" }
]
requires-python = ">=3.14"
dependencies = [
    "inquirer>=3.4.1",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "youtube-transcript-api>=1.2.3",
    "yt-dlp>=2025.12.8",
]

[project.scripts]
playlist-summarizer = "playlist_summarizer:main"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"
